update Rust gen to use .wit spec
This commit updates the regeneration script for the Rust bindings to
use `wasi-nn.wit` types instead of `wasi-nn.witx`.

The scripts has also been run and the bindings (src/generated.rs) have
been updated.

This is currently being linked to the main version of wasmtime and is
expected to work with release 20.0.0 when it is released.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>